Add server routing to the server commands (#324)
Add a way to specify which server to route the `server` requests to by
using the mechanism of server routing.

Now, to send a request which is automatically routed to the most optimal
(currently, the primary) server, use:
```
server list
```

To send a request to the specific replica, use:
```
server list 127.0.0.1:21729
```
